wip: OcrQueueScreen
* Migrate to new work manager based OcrQueueJob

* Remove legacy OcrQueue

+++ app/src/main/java/xyz/sevive/arcaeaoffline/ui/ocr/queue/OcrQueueList.kt
@@ -12,9 +12,9 @@ import xyz.sevive.arcaeaoffline.core.database.entities.PlayResult
 @Composable
 fun OcrQueueList(
     uiItems: List<OcrQueueScreenViewModel.TaskUiItem>,
-    onSaveScore: (Int) -> Unit,
-    onDeleteTask: (Int) -> Unit,
-    onEditPlayResult: (Int, PlayResult) -> Unit,
+    onSaveScore: (Long) -> Unit,
+    onDeleteTask: (Long) -> Unit,
+    onEditPlayResult: (Long, PlayResult) -> Unit,
 ) {
     LazyColumn(verticalArrangement = Arrangement.spacedBy(dimensionResource(R.dimen.list_padding))) {
         items(uiItems, key = { it.id }) { uiItem ->
